move and rename binding
