+++ src/core/mod.rs
@@ -6,7 +6,6 @@ mod interval;
 mod offset_time;
 mod port;
 mod signature;
-mod time_rep;
 mod width_rep;
 
 pub use component::{Component, Namespace};
@@ -16,7 +15,6 @@ pub use control::{
 };
 pub use id::Id;
 pub use interval::Range;
-pub use offset_time::{Time, TimeSum};
+pub use offset_time::{Time, TimeSub, TimeSum};
 pub use port::{Expr, InterfaceDef, PortDef};
 pub use signature::{EventBind, Signature};
-pub use time_rep::{Binding, TimeSub};
