+++ src/utils/mod.rs
@@ -4,7 +4,7 @@ mod position;
 mod post_order;
 mod solver;
 
-pub use bind_map::BindMap;
+pub use bind_map::Binding;
 pub use position::{FileIdx, GPosIdx, GlobalPositionTable, PosData};
 pub use post_order::PostOrder;
 pub use solver::{FilSolver, SExp, ShareConstraints};
